improve error semantics on simulation route crash; make fins and nosecone mandatory parameters
