+++ lib/controllers/environment.py
@@ -264,7 +264,7 @@ class EnvController:
             logger.error(f"controllers.environment.simulate: {exc_str}")
             raise HTTPException(
                 status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail=f"Failed to simulate environment: {exc_str}",
+                detail=f"Failed to simulate environment, parameters may contain data that is not physically coherent: {exc_str}",
             ) from e
         else:
             return env_summary
