+++ lib/controllers/rocket.py
@@ -279,7 +279,7 @@ class RocketController:
             logger.error(f"controllers.rocket.simulate: {exc_str}")
             raise HTTPException(
                 status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail=f"Failed to simulate rocket: {exc_str}",
+                detail=f"Failed to simulate rocket, parameters may contain data that is not physically coherent: {exc_str}",
             ) from e
         else:
             return rocket_summary
