+++ requirements-dev.txt
@@ -1,3 +1,5 @@
 flake8
 pylint
 ruff
+pytest
+httpx
